fix some of the examples
