+++ auto_forward_example_test.go
@@ -1,0 +1,78 @@
+package servicebus_test
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"time"
+
+	"github.com/Azure/azure-service-bus-go"
+)
+
+type MessagePrinter struct{}
+
+func (mp MessagePrinter) Handle(ctx context.Context, msg *servicebus.Message) error {
+	fmt.Println(string(msg.Data))
+	return msg.Complete(ctx)
+}
+
+func Example_autoForward() {
+	ctx, cancel := context.WithTimeout(context.Background(), 40*time.Second)
+	defer cancel()
+
+	connStr := os.Getenv("SERVICEBUS_CONNECTION_STRING")
+	if connStr == "" {
+		fmt.Println("FATAL: expected environment variable SERVICEBUS_CONNECTION_STRING not set")
+		return
+	}
+
+	// Create a client to communicate with a Service Bus Namespace.
+	ns, err := servicebus.NewNamespace(servicebus.NamespaceWithConnectionString(connStr))
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	qm := ns.NewQueueManager()
+	target, err := ensureQueue(ctx, qm, "AutoForwardTargetQueue")
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	source, err := ensureQueue(ctx, qm, "AutoForwardSourceQueue", servicebus.QueueEntityWithAutoForward(target))
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	sourceQueue, err := ns.NewQueue(source.Name)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer func() {
+		_ = sourceQueue.Close(ctx)
+	}()
+
+	if err := sourceQueue.Send(ctx, servicebus.NewMessageFromString("forward me to target!")); err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	targetQueue, err := ns.NewQueue(target.Name)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer func() {
+		_ = targetQueue.Close(ctx)
+	}()
+
+	if err := targetQueue.ReceiveOne(ctx, MessagePrinter{}); err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// Output: forward me to target!
+}
